# openapi/api-v1.yaml
openapi: 3.0.3
info:
  title: EsquiloSpeak API
  description: API specifications for the EsquiloSpeak MVP language learning app.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local API Gateway
security:
  - BearerAuth: []
paths:
...
  /courses/{courseId}/lessons/{lessonId}/complete:
    post:
      summary: Mark lesson as completed
      description: Validates and completes a lesson for a user. Only allowed when all unique questions in the lesson have been answered correctly on the server.
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
          example: "en_for_vi"
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
          example: "lesson_101"
      responses:
        '200':
          description: Lesson marked as completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Lesson or course not found
          $ref: '#/components/responses/ErrorResponse'
        '409':
          description: Lesson has no valid questions (empty content)
          $ref: '#/components/responses/ErrorResponse'
        '422':
          description: Learner has not answered all unique questions correctly on the server
          $ref: '#/components/responses/ErrorResponse'
...
  /courses/{courseId}/reviews/due:
    get:
      summary: Get due reviews
      description: Fetch list of items (vocabulary/sentences) scheduled for review today.
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
          example: "en_for_vi"
      responses:
        '200':
          description: Due reviews retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    reviewItemId:
                      type: string
                      example: "rev_item_001"
                    type:
                      type: string
                      enum: [vocabulary, sentence, grammar_point, listening_item]
                      example: "vocabulary"
                    concept:
                      type: string
                      example: "Hello"
                    hint:
                      type: string
                      example: "Lời chào hỏi ban đầu khi gặp mặt"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /courses/{courseId}/review-attempts:
    post:
      summary: Submit review results
      description: Updates scheduling parameters (interval, ease factor) based on user performance quality.
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
          example: "en_for_vi"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reviewItemId
                - rating
              properties:
                reviewItemId:
                  type: string
                  example: "rev_item_001"
                rating:
                  type: string
                  enum: [again, hard, good, easy]
                  example: "good"
                responseTimeMs:
                  type: integer
                  example: 3200
      responses:
        '200':
          description: Review result updated
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - nextReviewAt
                properties:
                  success:
                    type: boolean
                    example: true
                  nextReviewAt:
                    type: string
                    format: date-time
                    example: "2026-06-27T10:00:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /courses/{courseId}/progress/summary:
    get:
      summary: Get user progress summary
      description: Returns learning progress details like completion rates and streak metrics.
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
          example: "en_for_vi"
      responses:
        '200':
          description: Progress statistics
          content:
            application/json:
              schema:
                type: object
                required:
                  - completedLessonsCount
                  - totalLessonsCount
                  - courseCompletionPercent
                  - streak
                  - accuracy
                  - learnedWordsCount
                  - dueReviewCount
                properties:
                  completedLessonsCount:
                    type: integer
                    format: int32
                    example: 12
                  totalLessonsCount:
                    type: integer
                    format: int32
                    example: 20
                  courseCompletionPercent:
                    type: number
                    format: double
                    example: 60.0
                  streak:
                    type: integer
                    format: int32
                    example: 5
                  accuracy:
                    type: number
                    format: double
                    example: 85.5
                  learnedWordsCount:
                    type: integer
                    format: int64
                    example: 45
                  dueReviewCount:
                    type: integer
                    format: int32
                    example: 8
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
...
  responses:
    ErrorResponse:
      description: Standard Error Response
      content:
        application/json:
          schema:
            type: object
            required:
              - error
              - meta
            properties:
              error:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    example: "BAD_REQUEST"
                  message:
                    type: string
                    example: "Invalid input values"
                  details:
                    type: object
                    example: {}
              meta:
                type: object
                required:
                  - requestId
                  - apiVersion
                properties:
                  requestId:
                    type: string
                    example: "req_xyz987"
                  apiVersion:
                    type: string
                    example: "v1"
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            required:
              - error
              - meta
            properties:
              error:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    example: "UNAUTHORIZED"
                  message:
                    type: string
                    example: "Token is expired or invalid"
              meta:
                type: object
                required:
                  - requestId
                  - apiVersion
                properties:
                  requestId:
                    type: string
                    example: "req_xyz987"
                  apiVersion:
                    type: string
                    example: "v1"
    ForbiddenError:
      description: Authenticated user does not have permission to access this resource
      content:
        application/json:
          schema:
            type: object
            required:
              - error
              - meta
            properties:
              error:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    example: "FORBIDDEN"
                  message:
                    type: string
                    example: "Access denied"
              meta:
                type: object
                required:
                  - requestId
                  - apiVersion
                properties:
                  requestId:
                    type: string
                    example: "req_xyz987"
                  apiVersion:
                    type: string
                    example: "v1"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Nhập JWT access token để gọi các API được bảo vệ.